requirements.txt
blinker==1.9.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
reportlab==4.4.4
Werkzeug==3.1.3
